requirements.txt
requests
pytest
openai
anthropic
google-genai